+++ docs/README.md
@@ -1,0 +1,3 @@
+# pdoc.dev
+
+This directory contains the source code for `pdoc.dev`. Run `make.py` to build the `index.html` file!
