+++ docs/index.jinja2
@@ -1,0 +1,182 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <!-- ⚠️ this file is autogenerated from index.jinja2, do not edit manually. -->
+
+    <title>pdoc 📚</title>
+    <link rel="icon" type="image/svg+xml"
+          href="data:image/svg+xml,{% filter urlencode %}{% include "favicon.svg" %}{% endfilter %}"/>
+    <style type="text/css">{% include 'resources/bootstrap.min.css' %}</style>
+    {% filter replace("\n","") | replace("    ","") %}
+        <style type="text/css">{{ pygments_css }}</style>
+        <style type="text/css">
+
+            .pdoc-logo {
+                width: 200px;
+            }
+
+            {% set boxwidth = 450 %}
+            {% set boxpad = 20 %}
+            {% set minw = boxwidth * 2 + boxpad * 4 %}
+
+            @media (min-width: {{ minw }}px) {
+                #example > input, #example > label {
+                    display: none;
+                }
+            }
+
+            @media (max-width: {{ minw - 1 }}px) {
+                #example-code, #example-render {
+                    display: none;
+                }
+
+                #view-code:checked ~ #example-code {
+                    display: block;
+                }
+
+                #view-render:checked ~ #example-render {
+                    display: block;
+                }
+            }
+
+            #example > label {
+                margin-bottom: 1rem;
+            }
+
+            #example > label[for=view-code] {
+                border-top-right-radius: 0;
+                border-bottom-right-radius: 0;
+            }
+
+            #example > label[for=view-render] {
+                border-top-left-radius: 0;
+                border-bottom-left-radius: 0;
+            }
+
+            #example-invocation {
+                background-image: url("data:image/svg+xml,{% filter urlencode %}{% include "resources/arrow.svg" %}{% endfilter %}");
+                background-repeat: no-repeat;
+                background-position: bottom center;
+            }
+
+            #example-invocation code {
+                font-size: 1.2em;
+                display: inline-block;
+                border: solid var(--bs-dark) 1px;
+                background-color: rgba(255, 255, 255, 0.8);
+                color: var(--bs-gray-dark);
+                border-radius: 3px;
+                padding: 3px 7px;
+                margin-bottom: 32px;
+            }
+
+            #example-render, #example-code {
+                width: {{ boxwidth }}px;
+                max-width: 98vw;
+                min-height: 350px;
+                margin: 0 {{ boxpad }}px;
+                flex: 0 0 auto;
+                box-shadow: rgba(0, 0, 0, 0.15) 0 20px 68px;
+                border-radius: 5px;
+            }
+
+            #example-code {
+                background-color: rgb(253, 246, 227) !important;
+                position: relative;
+                overflow: auto;
+            }
+
+            #example-code svg {
+                position: absolute;
+                left: 14px;
+                top: 10px;
+            }
+
+            #example-code .title {
+                margin-top: 5px;
+                text-align: center;
+            }
+
+            #example-code .highlight {
+                margin: 15px;
+            }
+
+            #example-code pre {
+                overflow: inherit;
+            }
+
+        </style>
+    {% endfilter %}
+    <script src="/.devd.livereload.js"></script>
+</head>
+<body class="my-4">
+<div class="container-fluid">
+    <div class="row py-4 my-4 align-items-center">
+        <div class="col-md-6 text-center text-md-end">{% include "logo.svg" %}</div>
+        <div class="col-md-6 pt-md-0 pt-4 text-center text-md-start"><h3>API Documentation<br>for Python Projects</h3>
+        </div>
+    </div>
+</div>
+
+<div id="example" class="d-flex align-items-stretch justify-content-center flex-wrap">
+    <div id="example-invocation" class="col-12 text-center">
+        <code class="shadow">pdoc demo.py</code>
+    </div>
+    <input type="radio" class="btn-check" name="view-selector" id="view-code" autocomplete="off" checked>
+    <label class="btn btn-outline-success" for="view-code">Source Code</label>
+    <input type="radio" class="btn-check" name="view-selector" id="view-render" autocomplete="off">
+    <label class="btn btn-outline-success" for="view-render">Rendered Docs</label>
+    <div class="col-12"></div>
+    <div id="example-code">
+        <svg xmlns="http://www.w3.org/2000/svg" width="54" height="14" viewBox="0 0 54 14">
+            <g>
+                <circle cx="6" cy="6" r="6" fill="#FF5F56" stroke="#E0443E" stroke-width=".5"></circle>
+                <circle cx="26" cy="6" r="6" fill="#FFBD2E" stroke="#DEA123" stroke-width=".5"></circle>
+                <circle cx="46" cy="6" r="6" fill="#27C93F" stroke="#1AAB29" stroke-width=".5"></circle>
+            </g>
+        </svg>
+        <div class="title">demo.py</div>
+        {{ example_html }}
+    </div>
+    <iframe id="example-render" src="docs/demo.html"></iframe>
+</div>
+
+<div class="text-center my-4 py-4">
+    <h3>Installation</h3>
+    <div class="d-inline-block" style="width: 16rem">
+        <div class="input-group input-group-lg shadow">
+            <span class="input-group-text">
+                {% include "resources/terminal-fill.svg" %}
+            </span>
+            <input type="text" class="form-control" value="pip install pdoc">
+        </div>
+    </div>
+    <div class="mb-3 text-muted">Latest Release: {{ __version__ }}</div>
+    <a href="/docs/pdoc.html" class="btn btn-lg btn-success shadow mb-4 d-inline-flex align-items-center">
+        {% include "resources/book-half.svg" %}
+        &nbsp;Documentation
+    </a>
+    <a href="https://github.com/mitmproxy/pdoc/blob/main/CHANGELOG.md"
+       class="btn btn-lg btn-danger shadow mb-4 d-inline-flex align-items-center">
+        {% include "resources/newspaper.svg" %}
+        &nbsp;Changelog
+    </a>
+    <div class="d-md-none"></div>
+    <a href="https://pypi.org/project/pdoc/"
+       class="btn btn-lg btn-warning shadow mb-4 d-inline-flex align-items-center">
+        {% include "resources/box-seam.svg" %}
+        &nbsp;PyPI
+    </a>
+    <a href="https://github.com/mitmproxy/pdoc"
+       class="btn btn-lg btn-dark shadow mb-4 d-inline-flex align-items-center">
+        {% include "resources/github.svg" %}
+        &nbsp;GitHub
+    </a>
+</div>
+
+<div class="d-none d-md-block">{% include "resources/github-button.html" %}</div>
+</body>
+</html>
