+++ docs/modd.conf
@@ -1,0 +1,7 @@
+# https://github.com/cortesi/modd
+# cd .. && modd -f docs/modd.conf
+
+pdoc/** test/** docs/index.jinja2 {
+    prep: python docs/make.py
+    daemon: devd -m -p 8000 docs
+}
