+++ pdoc/templates/html_module.jinja2
@@ -7,7 +7,7 @@
         /* General Styles */
 
         :root {
-            --main-pad: 3rem;
+            --main-pad: calc(min(3rem, 3vw));
         }
 
         html, main {
@@ -51,6 +51,10 @@
                 background-repeat: no-repeat;
                 background-position: center center;
                 background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 30 30'%3e%3cpath stroke='rgba%280, 0, 0, 0.55%29' stroke-linecap='round' stroke-miterlimit='10' stroke-width='2' d='M4 7h22M4 15h22M4 23h22'/%3e%3c/svg%3e");
+            }
+
+            main {
+                padding-top: 2rem !important;
             }
 
             label[for=navtoggle]:hover {
@@ -201,6 +205,7 @@
         }
 
         .modulename {
+            margin-top: 0;
             font-weight: bold;
         }
 
@@ -239,7 +244,7 @@
         }
 
         details > div {
-            margin-top: calc(0px - var(--shift));
+            margin-top: calc(0px - var(--shift) / 2);
             text-align: left;
         }
 
@@ -377,7 +382,7 @@
     {{- decorators(fn) -}}
     {% if fn.name == "__init__" %}
         <span class="name">{{ ".".join(fn.qualname.split(".")[:-1]) }}</span>
-        {{ sig | replace("(self)","()") | replace("(self, ","(") }}
+        {{- sig | replace("(self)","()") | replace("(self, ","(") -}}
     {% else %}
         <span class="def">{{ fn.funcdef }}</span>
         <span class="name">{{ fn.name }}</span>{{ sig }}:
@@ -468,8 +473,8 @@
                         <a href="./{{ "../" * loop.revindex0 }}{{ part }}.html">{{ modulename }}</a>
                         {%- if loop.nextitem %}.{% endif %}
                     {%- endfor %}</h1>
-                {{ view_source(module) }}
                 {{ docstring(module) }}
+                {{ view_source(module) }}
             </section>
 
             {% for m in module.flattened_members %}
@@ -526,6 +531,7 @@
     </div>
     {% if mtime %}
         <script>
+            /* Periodically check with the pdoc server if there have been any changes. */
             let mtime_generated = {{ mtime | tojson }},
                 url = location.href;
             url += (url.includes("?") ? "&" : "?") + "mtime=1";
