+++ test/modd.conf
@@ -1,3 +1,6 @@
+# https://github.com/cortesi/modd
+# cd .. && modd -f test/modd.conf
+
 pdoc/** test/** {
     prep: pdoc --html -o .render test/snapshots/demo.py pdoc --edit-on-github demo=https://example.com/demo
     daemon: devd -m -p 9000 .render
