+++ test/test_main.py
@@ -10,7 +10,7 @@ here = Path(__file__).parent
 
 
 def test_cli(tmp_path):
-    cli(args=[str(here / "snapshots" / "demo.py"), "-o", str(tmp_path)])
+    cli(args=[str(here / "snapshots" / "demo_long.py"), "-o", str(tmp_path)])
     assert (tmp_path / "demo.html").read_text().startswith("<!doctype html>")
     assert (tmp_path / "index.html").read_text().startswith("<!doctype html>")
 
@@ -21,15 +21,15 @@ def test_cli_web(monkeypatch):
             "pdoc.web.DocServer.serve_forever", side_effect=KeyboardInterrupt
         ) as serve_forever:
             with pytest.raises(KeyboardInterrupt):
-                cli(args=[str(here / "snapshots" / "demo.py")])
+                cli(args=[str(here / "snapshots" / "demo_long.py")])
             assert open_browser.call_args == call("http://localhost:8080/demo")
             assert serve_forever.call_args == call()
 
 
 def test_api(tmp_path):
-    assert pdoc(here / "snapshots" / "demo.py").startswith("<!doctype html>")
+    assert pdoc(here / "snapshots" / "demo_long.py").startswith("<!doctype html>")
     with pytest.raises(ValueError, match="Invalid rendering format"):
-        assert pdoc(here / "snapshots" / "demo.py", format="invalid")
+        assert pdoc(here / "snapshots" / "demo_long.py", format="invalid")
     with pytest.raises(ValueError, match="No valid module specifications"):
         with pytest.warns(RuntimeWarning, match="Cannot find spec"):
             assert pdoc(
